+++ src/leiningen/ezbake_core.clj
@@ -34,23 +34,6 @@
     (deputils/cp-files-from-jar jar-entries jar-file resource-path)))
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;;; Resource Directory Type Helpers
-(defn copy-dir [src-dir dest-dir]
-  (let [src (io/file src-dir)
-        dest (io/file dest-dir)]
-    (if (.exists src)
-      (if (.isDirectory src)
-        (do
-          (.mkdirs dest)
-          (doseq [file (.listFiles src)]
-            (copy-dir (.getPath file) (str (io/file dest (.getName file))))))
-        (do
-          (lein-main/info (format "Copying lein-ezbake resources from included directory: %s to %s."
-                                        src-dir dest-dir))
-          (io/copy src dest)))
-      (throw (RuntimeException. (format "Resource directory %s does not exist." (str src-dir)))))))
-
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;;; Consumable API
 
 (defn prepare-resource-dir
@@ -63,15 +46,9 @@
   (let [template-type (get-in project [:lein-ezbake
                                        :resources
                                        :type]
-                              :jar)
-        include-resource-dir (get-in project [:lein-ezbake
-                                              :resources
-                                              :include-dir]
-                                     nil)]
+                              :jar)]
     (case template-type
       :git (throw (RuntimeException.
                    (format "Resource type, %s, not implemented."
                            (str template-type))))
-      :jar (copy-jar-resources core/resource-prefix core/resource-path))
-    (when (some? include-resource-dir)
-      (copy-dir include-resource-dir core/resource-path))))
+      :jar (copy-jar-resources core/resource-prefix core/resource-path))))